fix: enhance toggle explore pane performance
